fix: for setp 4 in onboarding in trader page scroll to work on whole page
