+++ packages/trader/src/AppV2/Components/OnboardingGuide/guide-container.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import Joyride, { CallBackProps, STATUS } from 'react-joyride';
 import GuideTooltip from './guide-tooltip';
 import STEPS from './steps-config';
+import { useSwipeable } from 'react-swipeable';
 
 type TGuideContainerProps = {
     should_run: boolean;
@@ -20,39 +21,53 @@ const GuideContainer = ({ should_run, onFinishGuide }: TGuideContainerProps) => 
         if (finished_statuses.includes(status)) onFinishGuide();
     };
 
+    const swipe_handlers = useSwipeable({
+        onSwipedUp: () => {
+            if (STEPS[step_index].title === 'scroll-icon') {
+                document.querySelector('.trade__chart')?.scrollIntoView();
+                setStepIndex((prev: number) => prev + 1);
+            }
+        },
+        preventDefaultTouchmoveEvent: true,
+        trackTouch: true,
+        trackMouse: true,
+    });
+
     return (
-        <Joyride
-            continuous
-            callback={callbackHandle}
-            disableCloseOnEsc
-            disableOverlayClose
-            disableScrolling
-            floaterProps={{
-                styles: {
-                    arrow: {
-                        length: 4,
-                        spread: 8,
-                        display: step_index === 3 ? 'none' : 'inline-flex',
+        <div {...swipe_handlers}>
+            <Joyride
+                continuous
+                callback={callbackHandle}
+                disableCloseOnEsc
+                disableOverlayClose
+                disableScrolling
+                floaterProps={{
+                    styles: {
+                        arrow: {
+                            length: 4,
+                            spread: 8,
+                            display: step_index === 3 ? 'none' : 'inline-flex',
+                        },
                     },
-                },
-            }}
-            run={should_run}
-            showSkipButton
-            steps={STEPS}
-            spotlightPadding={0}
-            scrollToFirstStep
-            styles={{
-                options: {
-                    arrowColor: 'var(--component-textIcon-normal-prominent)',
-                    overlayColor: 'var(--core-color-opacity-black-600)',
-                },
-                spotlight: {
-                    borderRadius: 'unset',
-                },
-            }}
-            stepIndex={step_index}
-            tooltipComponent={props => <GuideTooltip {...props} setStepIndex={setStepIndex} />}
-        />
+                }}
+                run={should_run}
+                showSkipButton
+                steps={STEPS}
+                spotlightPadding={0}
+                scrollToFirstStep
+                styles={{
+                    options: {
+                        arrowColor: 'var(--component-textIcon-normal-prominent)',
+                        overlayColor: 'var(--core-color-opacity-black-600)',
+                    },
+                    spotlight: {
+                        borderRadius: 'unset',
+                    },
+                }}
+                stepIndex={step_index}
+                tooltipComponent={props => <GuideTooltip {...props} setStepIndex={setStepIndex} />}
+            />
+        </div>
     );
 };
